refactor(cloudformation): extract inline ZipFile packaging
CloudFormationResourceProvisioner is a monolith being dismantled, so the
cfn-response bundling added there moves to its own InlineZipPackager,
ready to travel with the eventual Lambda provisioner extraction. No
behavior change.

(cherry picked from commit d0d4bff84bda9c143606a8922500515dbe5bc0c0)
